move indexer to its own class, kill vault
