+++ config/application.rb
@@ -34,11 +34,11 @@ module Gemcutter
 
     config.after_initialize do
       Hostess.local = $rubygems_config[:local_storage]
-
-      YAML::ENGINE.yamler = "syck"
     end
 
     config.plugins = [:dynamic_form]
     config.plugins << :heroku_asset_cacher if $rubygems_config[:asset_cacher]
+
+    config.autoload_paths << "./app/jobs"
   end
 end
