+++ lib/tasks/gemcutter.rake
@@ -43,64 +43,9 @@ namespace :gemcutter do
         b.report("   pre index") { g.stringify(g.prerelease_index) }
       end
     end
-
-    desc "fix the index"
-    task :reprocess => :environment do
-      index = Gem::SourceIndex.new
-
-      Rubygem.with_versions.each do |rubygem|
-        rubygem.versions.each do |version|
-
-          install = "#{rubygem.name}-#{version.number}"
-          quick_path = "quick/Marshal.#{Gem.marshal_version}/#{install}.gemspec.rz"
-
-          if VaultObject.exists?(quick_path)
-            puts ">> Processing #{install}"
-            begin
-              spec = Marshal.load(Gem.inflate(VaultObject.value(quick_path)))
-            rescue Exception => e
-              puts ">> EXCEPTION: #{e}"
-              version.update_attribute(:indexed, false)
-            end
-
-            version.description = spec.description
-            version.summary = spec.summary
-            version.number = spec.version.to_s
-
-            platform = spec.original_platform
-            platform = Gem::Platform::RUBY if platform.nil? or platform.empty?
-            version.platform = platform
-            version.save
-
-            spec.development_dependencies.each { |dep| version.dependencies.create_from_gem_dependency!(dep) }
-
-            index.add_spec(spec)
-          else
-            puts ">> BAD GEM: #{install}"
-            version.update_attribute(:indexed, false)
-          end
-        end
-      end
-
-      puts ">> ding, gems are done!"
-      File.open("/tmp/index", "wb") { |f| f.write Marshal.dump(index) }
-    end
   end
 
   namespace :import do
-    desc 'Make sure all of the gems are on S3'
-    task :verify => :environment do
-      return unless Rails.env.production?
-      Version.all.each do |version|
-        path = "#{version.rubygem.name}-#{version.number}.gem"
-        gem_path = "gems/#{path}"
-        spec_path = "quick/Marshal.4.8/#{path}spec.rz"
-
-        puts gem_path unless VaultObject.exists?(gem_path)
-        puts spec_path unless VaultObject.exists?(spec_path)
-      end
-    end
-
     desc 'Upload gems to s3 like a boss'
     task :upload => :environment do
       return unless Rails.env.production?
@@ -181,11 +126,7 @@ namespace :gemcutter do
         file.write new_content
       end
 
-      class Updater
-        include Vault
-      end
-
-      updater = Updater.new
+      updater = Indexer.new
       html    = Nokogiri.parse(open("http://rubyforge.org/frs/?group_id=126"))
       links   = html.css("a[href*='#{version}']").map { |n| n["href"] }
       links.each do |link|
