+++ test/unit/download_test.rb
@@ -1,4 +1,4 @@
-require File.dirname(__FILE__) + '/../test_helper'
+require 'test_helper'
 
 class DownloadTest < ActiveSupport::TestCase
   should "load up all downloads with just raw strings and process them" do
