+++ test/unit/helpers/rubygems_helper_test.rb
@@ -1,4 +1,4 @@
-require File.expand_path("../../../test_helper", __FILE__)
+require 'test_helper'
 
 class RubygemsHelperTest < ActionView::TestCase
   include Rails.application.routes.url_helpers
