+++ test/unit/linkset_test.rb
@@ -1,4 +1,4 @@
-require File.dirname(__FILE__) + '/../test_helper'
+require 'test_helper'
 
 class LinksetTest < ActiveSupport::TestCase
   should belong_to :rubygem
