+++ test/unit/pusher_test.rb
@@ -1,4 +1,4 @@
-require File.dirname(__FILE__) + '/../test_helper'
+require 'test_helper'
 
 class PusherTest < ActiveSupport::TestCase
   context "getting the server path" do
